requirements.txt
aiohttp
playwright>=1.40.0
pyodbc>=4.0.35
python-dotenv>=1.0.0
fake-useragent>=1.4.0
tqdm>=4.66.0
textblob>=0.17.0
langdetect>=1.0.9
transformers>=4.30.0
torch>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.4.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
anthropic
groq